+++ src/main/java/clientcore/websocket/models/notifications/ProjectGrantPermissionsNotification.java
@@ -8,10 +8,10 @@ public class ProjectGrantPermissionsNotification implements INotificationData {
     public final String grantUsername;
 
     @JsonProperty("PermissionLevel")
-    public final Byte permissionLevel;
+    public final int permissionLevel;
 
     public ProjectGrantPermissionsNotification(@JsonProperty("GrantUsername") String grantUsername,
-                                               @JsonProperty("PermissionLevel") Byte permissionLevel) {
+                                               @JsonProperty("PermissionLevel") int permissionLevel) {
         this.grantUsername = grantUsername;
         this.permissionLevel = permissionLevel;
     }
